requirements.txt
discord.py==2.3.2
numpy==1.25.2
pydub==0.25.1
pytube==15.0.0
Requests==2.31.0
youtube_search==2.1.2
